biltong/requirements.txt
asgiref==3.4.1
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Django==3.2.2
django-appconf==1.0.5
django-cors-headers==3.10.1
django-hosts==5.0
django-markdownfield==0.10.0
djangorestframework==3.12.4
drf-jwt==1.19.1
geojson==2.5.0
idna==3.3
importlib-metadata==4.9.0
Markdown==3.3.6
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pilkit==2.0
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
PyJWT==1.7.1
pyowm==3.2.0
pyparsing==3.0.6
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
zipp==3.6.0
